+++ docs/recipes/automation/cloud/bastion/setup-dp-gke-platformprovisioner.sh
@@ -255,7 +255,7 @@ function platform_provisioner_verify_access() {
 
     cd ${HOME}/platform-provisioner
     echo "########## Verify GCP Account Access ##########"
-    export PIPELINE_INPUT_RECIPE="docs/recipes/tests/test-gcp.yaml"
+    export PIPELINE_INPUT_RECIPE="tests/e2e/test-gcp.yaml"
     ./dev/platform-provisioner.sh 
     platform_provisioner_update_deploy_recipe